# Pipeline commands in Client#raw_push