@mperham 2019-07-27
Would it be better to port Ruby’s smarter MULTI behavior rather than hand roll it here?