@pgeraghty 2019-07-27
What's odd if you see the "monitor" log is that Ruby appears to send the "multi" command first too (there's a significant delay there that looks like a round trip), which is why I was looking for confirmation that I'm not on a wild goose chase.

I'm positive you're right about the approach of course; I'm just unsure of the various idiosyncrasies of Redis (re. transactions and pipelining), so I thought I'd put this out there for a response.

I'll certainly delve further into this.