@pgeraghty 2019-07-29
I've made a pull request to the [Crystal Redis shard](/stefanwille/crystal-redis) to address this at the core as you suggested; below are the benchmark results using [my branch](/pgeraghty/crystal-redis/tree/pipeline_multi) as AFTER.

### 10,000 Sidekiq job pushes to local Redis

**BEFORE**:
00:00:00.874722330 => ~875ms

**AFTER**:
00:00:00.510320163 => ~510ms


### 100 Sidekiq job pushes to remote Redis w/ ~45ms ping

**BEFORE**:
00:00:12.878749613 => ~12.9s

**AFTER**:
00:00:04.562689087 => ~4.5s


### Results of cloned benchmark using Ruby Redis 4.1.2 to compare

10,000 Sidekiq job pushes to local Redis: ~1.0s
100 Sidekiq job pushes to remote Redis w/ ~45ms ping: ~9.3s

The tables have turned, as Ruby is now in a position where it is making 2 round trips to Crystal's 1 (previously 3).