@hugopl 2021-03-17
The mentioned patch was merged ages ago https://github.com/stefanwille/crystal-redis/commit/d7b437a8922071e72148b262e179706148b4a8f1

I think this issue can be closed.